# openapi.yml
openapi: 3.0.3
info:
  title: Bulma API
  version: 1.0.0
  description: These are the OpenAPI specs that describe the REST API of Bulma.
  termsOfService: https://bulma.org/terms-and-conditions
  contact:
    name: Bulma
    url: https://bulma.org
    email: info@bulma.org
  license:
    name: GPL-3.0
    url: https://www.gnu.org/licenses/gpl-3.0.en.html
...
servers:
  - url: https://demo.bulma.org/index.php/api/v1/
  - url: http://localhost/index.php/api/v1/
...
paths:
...
  /providers/{providerId}:
    get:
      tags:
        - providers
      summary: Get a provider
      parameters:
        - name: providerId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderRecord'
        '401':
          description: Unauthorized
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - BearerToken: [ ]
        - BasicAuth: [ ]
    put:
      tags:
        - providers
      summary: Update a provider
      parameters:
        - name: providerId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProviderPayload'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderRecord'
        '401':
          description: Unauthorized
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-codegen-request-body-name: body
      security:
        - BearerToken: [ ]
        - BasicAuth: [ ]
    delete:
      tags:
        - providers
      summary: Delete a provider
      parameters:
        - name: providerId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - BearerToken: [ ]
        - BasicAuth: [ ]
...
components:
  schemas:
...
    ProviderRecord:
      type: object
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        mobile:
          type: string
        phone:
          type: string
        address:
          type: string
        city:
          type: string
        zip:
          type: string
        notes:
          type: string
        timezone:
          type: string
        language:
          type: string
        isPrivate:
          type: boolean
        ldapDn:
          type: string
        services:
          type: array
          items:
            type: integer
        settings:
          type: object
          properties:
            username:
              type: string
            password:
              type: string
            notifications:
              type: boolean
            calendarView:
              type: string
            googleSync:
              type: boolean
            googleCalendar:
              type: string
            googleToken:
              type: string
            caldavSync:
              type: boolean
            caldavUrl:
              type: string
            caldavUsername:
              type: string
            caldavPassword:
              type: string
            syncFutureDays:
              type: string
            syncPastDays:
              type: string
            workingPlan:
              type: string
      example:
        id: 1
        firstName: Chris
        lastName: Doe
        email: chris@example.org
        mobile: '+10000000000'
        phone: '+10000000000'
        address: Test Street 1A
        city: Some Place
        zip: '12345'
        notes: This is a test provider.
        timezone: UTC
        language: english
        services: [ ]
        isPrivate: false
        ldapDn: null
        settings:
          username: chrisdoe
          password: Password@123
          notifications: true
          calendarView: default
          googleSync: false
          googleCalendar: null
          googleToken: null
          caldavSync: false
          caldavUrl: null
          caldavUsername: null
          caldavPassword: null
          syncFutureDays: 90
          syncPastDays: 30
          workingPlan:
            sunday: null
            monday:
              start: '09:00'
              end: '17:00'
              breaks: [ ]
            tuesday:
              start: '09:00'
              end: '17:00'
              breaks: [ ]
            wednesday:
              start: '09:00'
              end: '17:00'
              breaks: [ ]
            thursday:
              start: '09:00'
              end: '17:00'
              breaks: [ ]
            friday:
              start: '09:00'
              end: '17:00'
              breaks: [ ]
            saturday: null
    ProviderPayload:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        mobile:
          type: string
        phone:
          type: string
        address:
          type: string
        city:
          type: string
        zip:
          type: string
        notes:
          type: string
        timezone:
          type: string
        language:
          type: string
        isPrivate:
          type: boolean
        ldapDn:
          type: string
        services:
          type: array
          items:
            type: integer
        settings:
          type: object
          properties:
            username:
              type: string
            password:
              type: string
              format: password
            notifications:
              type: boolean
            calendarView:
              type: string
            googleSync:
              type: boolean
            googleCalendar:
              type: string
            googleToken:
              type: string
            caldavSync:
              type: boolean
            caldavUrl:
              type: string
            caldavUsername:
              type: string
            caldavPassword:
              type: string
            syncFutureDays:
              type: string
            syncPastDays:
              type: string
            workingPlan:
              type: string
      example:
        firstName: Chris
        lastName: Doe
        email: chris@example.org
        mobile: '+10000000000'
        phone: '+10000000000'
        address: Test Street 1A
        city: Some Place
        zip: '12345'
        notes: This is a test provider.
        timezone: UTC
        language: english
        isPrivate: false
        ldapDn: null
        services: [ ]
        settings:
          username: chrisdoe
          password: Password@123
          notifications: true
          calendarView: default
          googleSync: false
          googleCalendar: null
          googleToken: null
          caldavSync: false
          caldavUrl: null
          caldavUsername: null
          caldavPassword: null
          syncFutureDays: 90
          syncPastDays: 30
          workingPlan:
            sunday: null
            monday:
              start: '09:00'
              end: '17:00'
              breaks: [ ]
            tuesday:
              start: '09:00'
              end: '17:00'
              breaks: [ ]
            wednesday:
              start: '09:00'
              end: '17:00'
              breaks: [ ]
            thursday:
              start: '09:00'
              end: '17:00'
              breaks: [ ]
            friday:
              start: '09:00'
              end: '17:00'
              breaks: [ ]
            saturday: null
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      example:
        success: false
        message: Test error message with debug info.
  securitySchemes:
    BearerToken:
      type: http
      scheme: bearer
      description: >-
        You can generate the API key from the settings page to authorize with
        Bearer Token.
    BasicAuth:
      type: http
      scheme: basic
      description: >-
        You can use the credentials of an admin user to authorize with Basic
        Auth.